Renamed step adding packages to project
